Large scale API changes and qa setup
There's a load of changes here in terms of how the API is structured. Additionally, there is now an abstraction implemented to allow QA to happen. This an abstraction layer on top of the db models which is customisable by the different parts of the system which want to have questions.