Fix bug with checking the auth on each page load
On first load the app often thought you weren't logged in because the route beforeEach completed before the refreshDetails call in the main.js. This commit fixes this by using an active check on each route navigation to ensure the user status is up to date. This also gives us the opportunity to confirm the details each hour.