Transition fixtures to a custom loading pattern
Fixtures are great and yay free stuff from django, but, it doesn't give us much flexibility in terms of hooking into what's changed so that we can notify users and it makes it harder for users to modify the policy model objects as they have to conform to the django fixtures structure. This is a first pass at changing that.