Switch to using relative refs rather than just names
Names are vague, also relative ids allow the inputting user to enter an id and reference that id without having to worry about what the actual underlying id in the db is. This was were the root of the problem was with having ids because it meant you'd have to keep track of the ids you've used and where outside of the context of the entity you're editing. Relative refs don't require that.