Accommodate different non-current-institution respondents in the pack views/components and add start and end to packs.
This is a bit of a large commit, but the change was quite far-reaching in the end.
Essentially, instead of always using the institution the user is attached to we need to be able to support different responders.
This is now implemented in the backend and this is the frontend work to make that happen.
Policy responses will always be associated with the institution the user is associated with but maturity responses need to be associated with whatever team the user wants.
Maturity users can't currently create their own teams, but that will come shortly.
This change also allows a future change where the user is associated with multiple institutions, rather than just one.

Additionally, this change accommodates another related change - finishing packs.
The pack flow now has a start and end screen, the start shows info about the questions, the end allows the user to declare the pack complete.